I moved all Util* classes to "packages/UtilAna".  I updated "OnlMonTrigNim", which now shows a meaningful quantity.
