+++ online/decoder_maindaq/CMakeLists.txt
@@ -31,7 +31,7 @@ set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++0x -I$ENV{MY_INSTALL}/include/ -
 set(CMAKE_C_FLAGS   "${CMAKE_C_FLAGS}   -DBIT64    -I$ENV{MY_INSTALL}/include/ -I$ENV{OFFLINE_MAIN}/include/")
 
 add_library(decoder_maindaq SHARED ${sources} ${dicts})
-target_link_libraries(decoder_maindaq -linterface_main -lfun4all -lphool -lgeom_svc)
+target_link_libraries(decoder_maindaq -linterface_main -lfun4all -lphool -lgeom_svc -lUtilAna)
 
 message(${CMAKE_PROJECT_NAME} " will be installed to " ${CMAKE_INSTALL_PREFIX})
