+++ online/decoder_maindaq/CodaInputManager.cc
@@ -4,8 +4,8 @@
 #include <cstdlib>
 #include <fstream>
 #include <unistd.h>
+#include <UtilAna/UtilOnline.h>
 #include "evio.h"
-#include "UtilOnline.h"
 #include "CodaInputManager.h"
 using namespace std;
