--- online/decoder_maindaq/UtilOnlineLinkDef.h
@@ -1,13 +1,0 @@
-#ifdef __CINT__
-
-#pragma link off all globals;
-#pragma link off all classes;
-#pragma link off all functions;
-
-#pragma link C++ class UtilOnline;
-//#pragma link C++ namespace UtilOnline;
-//#pragma link C++ defined_in namespace UtilOnline;
-//#pragma link C++ function UtilOnline::CodaFile2RunNum;
-//#pragma link C++ function UtilOnline::RunNum2CodaFile;
-
-#endif /* __CINT__ */
