+++ online/onlmonserver/LinkDef.h
@@ -3,8 +3,6 @@
 #pragma link off all function;
 #pragma link off all global;
 
-#pragma link C++ class UtilSQHit-!;
-
 #pragma link C++ class OnlMonServer-!;
 #pragma link C++ class OnlMonClient-!;
 #pragma link C++ class OnlMonUI-!;
