+++ online/onlmonserver/OnlMonProp.cc
@@ -10,9 +10,9 @@
 #include <phool/getClass.h>
 #include <geom_svc/GeomSvc.h>
 //#include <geom_svc/CalibParamInTimeTaiwan.h>
+#include <UtilAna/UtilHist.h>
 #include "OnlMonServer.h"
 #include "OnlMonProp.h"
-#include "UtilHist.h"
 using namespace std;
 
 OnlMonProp::OnlMonProp(const PropType_t type) : m_type(type)
@@ -120,7 +120,6 @@ int OnlMonProp::DrawMonitor()
     pad0->cd(pl+1);
     h1_ele[pl]->Draw();
   }
-  can0->AddMessage("OK");
   can0->SetStatus(OnlMonCanvas::OK);
 
   OnlMonCanvas* can1 = GetCanvas(1);
@@ -132,7 +131,6 @@ int OnlMonProp::DrawMonitor()
     UtilHist::AutoSetRange(h1_time[pl]);
     h1_time[pl]->Draw();
   }
-  can1->AddMessage("OK");
   can1->SetStatus(OnlMonCanvas::OK);
 
   return 0;
