+++ online/onlmonserver/OnlMonTrigSig.cc
@@ -8,9 +8,9 @@
 #include <phool/PHIODataNode.h>
 #include <phool/getClass.h>
 #include <geom_svc/GeomSvc.h>
+#include <UtilAna/UtilHist.h>
 #include "OnlMonServer.h"
 #include "OnlMonTrigSig.h"
-#include "UtilHist.h"
 using namespace std;
 
 // BeforeInhNIM
@@ -148,7 +148,6 @@ int OnlMonTrigSig::DrawMonitor()
   pad0->Divide(1, 2);
   pad0->cd(1);  DrawTH2WithPeakPos(h2_bi_fpga);
   pad0->cd(2);  DrawTH2WithPeakPos(h2_ai_fpga);
-  can0->AddMessage("OK");
   can0->SetStatus(OnlMonCanvas::OK);
 
   OnlMonCanvas* can1 = GetCanvas(1);
@@ -157,7 +156,6 @@ int OnlMonTrigSig::DrawMonitor()
   pad1->Divide(1, 2);
   pad1->cd(1);  DrawTH2WithPeakPos(h2_bi_nim);
   pad1->cd(2);  DrawTH2WithPeakPos(h2_ai_nim);
-  can1->AddMessage("OK");
   can1->SetStatus(OnlMonCanvas::OK);
 
   OnlMonCanvas* can2 = GetCanvas(2);
@@ -166,7 +164,6 @@ int OnlMonTrigSig::DrawMonitor()
   pad2->Divide(1, 2);
   pad2->cd(1);  DrawTH2WithPeakPos(h2_rf  );
   pad2->cd(2);  DrawTH2WithPeakPos(h2_stop);
-  can2->AddMessage("OK");
   can2->SetStatus(OnlMonCanvas::OK);
 
   return 0;
