+++ packages/UtilAna/CMakeLists.txt
@@ -3,8 +3,19 @@ cmake_minimum_required(VERSION 2.6 FATAL_ERROR)
 project(UtilAna CXX C)
 
 # source code
-file(GLOB sources ${PROJECT_SOURCE_DIR}/*.cc)
-file(GLOB headers ${PROJECT_SOURCE_DIR}/*.h)
+file(GLOB          sources ${PROJECT_SOURCE_DIR}/*.cc)
+file(GLOB     dist_headers ${PROJECT_SOURCE_DIR}/*.h)
+file(GLOB non_dist_headers ${PROJECT_SOURCE_DIR}/LinkDef.h)
+list(REMOVE_ITEM dist_headers ${non_dist_headers})
+
+# ROOT dict generation
+add_custom_command (
+  OUTPUT UtilAna_Dict.cc
+  COMMAND rootcint
+  ARGS -f UtilAna_Dict.cc -noIncludePaths -inlineInputHeader -c -p
+    -I${PROJECT_SOURCE_DIR}/ -I${ROOT_PREFIX}/include/
+    ${dist_headers} ${non_dist_headers}
+)
 
 # ROOT
 find_program(ROOTCONF "root-config")
@@ -19,20 +30,16 @@ execute_process(COMMAND root-config --libs   OUTPUT_VARIABLE ROOT_LINK    OUTPUT
 
 set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++0x ${ROOT_CFLAGS}")
 
-add_library(UtilAna SHARED ${sources})
+add_library(UtilAna SHARED ${sources} UtilAna_Dict.cc)
 target_link_libraries(UtilAna -linterface_main -lgeom_svc)
 
 install(TARGETS UtilAna DESTINATION ${CMAKE_INSTALL_PREFIX}/lib)
-
-file(GLOB dist_headers ${PROJECT_SOURCE_DIR}/*.h)
-#file(GLOB non_dist_headers ${PROJECT_SOURCE_DIR}/LinkDef.h)
-#list(REMOVE_ITEM dist_headers ${non_dist_headers})
 install(FILES ${dist_headers} DESTINATION ${CMAKE_INSTALL_PREFIX}/include/${CMAKE_PROJECT_NAME}/)
 
 # Install the pcm files in case of ROOT 6.
-#execute_process(COMMAND root-config --version OUTPUT_VARIABLE ROOT_VER)
-#string(SUBSTRING ${ROOT_VER} 0 1 ROOT_VER)
-#if (ROOT_VER GREATER 5)
-#   add_custom_target(install_pcm ALL COMMAND mkdir -p ${CMAKE_INSTALL_PREFIX}/lib COMMAND cp -up *_rdict.pcm ${CMAKE_INSTALL_PREFIX}/lib)
-#   add_dependencies(install_pcm UtilAna)
-#endif()
+execute_process(COMMAND root-config --version OUTPUT_VARIABLE ROOT_VER)
+string(SUBSTRING ${ROOT_VER} 0 1 ROOT_VER)
+if (ROOT_VER GREATER 5)
+   add_custom_target(install_pcm ALL COMMAND mkdir -p ${CMAKE_INSTALL_PREFIX}/lib COMMAND cp -up *_rdict.pcm ${CMAKE_INSTALL_PREFIX}/lib)
+   add_dependencies(install_pcm UtilAna)
+endif()
