+++ packages/UtilAna/LinkDef.h
@@ -1,0 +1,9 @@
+#ifdef __CINT__
+#pragma link off all class;
+#pragma link off all function;
+#pragma link off all global;
+
+#pragma link C++ class UtilOnline-!;
+#pragma link C++ class UtilSQHit-!;
+
+#endif
